evaluation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
astunparse=1.6.3=pypi_0
audioread=3.0.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
coloredlogs=15.0.1=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.2=py39h774aba0_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
flatbuffers=23.5.26=pypi_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.10.0=pypi_0
gast=0.4.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
gpustat=1.1.1=pypi_0
grpcio=1.57.0=pypi_0
h5py=3.9.0=pypi_0
huggingface-hub=0.19.4=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=py39h06a4308_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h5eee18b_1
julius=0.2.7=pypi_0
keras=2.2.4=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
keras-self-attention=0.49.0=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=16.0.6=pypi_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.1.12=0
libcurand=10.3.3.129=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
librosa=0.10.1=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
llvmlite=0.40.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
mat4py=0.5.0=pypi_0
matplotlib=3.8.1=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
mock=5.1.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
msgpack=1.0.5=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numba=0.57.1=pypi_0
numpy=1.24.3=pypi_0
nvidia-ml-py=12.535.77=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.16.3=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pesq=0.0.4=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
platformdirs=3.10.0=pypi_0
pooch=1.7.0=pypi_0
protobuf=4.24.2=pypi_0
psutil=5.9.0=py39h5eee18b_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.3.1=py39_torch_2.0.0_cu118
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
pystoi=0.3.3=pypi_0
python=3.9.16=h7a1cb2a_2
pytorch=2.0.1=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cluster=1.6.1=py39_torch_2.0.0_cu118
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rir-generator=0.2.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.0=pypi_0
scikit-learn=1.3.0=py39h1128e8f_0
scipy=1.11.1=py39h5f9d8c6_0
seaborn=0.13.0=pypi_0
setuptools=67.8.0=py39h06a4308_0
six=1.16.0=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.6=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.13.0=pypi_0
tensorflow-estimator=1.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.33.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
timm=0.4.5=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.0=pypi_0
torch-stoi=0.1.2=pypi_0
torch-tb-profiler=0.4.1=pypi_0
torchaudio=2.0.2=py39_cu118
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu118
tqdm=4.65.0=py39hb070fc8_0
transformers=4.35.2=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=py39h06a4308_0
wcwidth=0.2.6=pypi_0
werkzeug=2.3.7=pypi_0
wget=3.2=pypi_0
wheel=0.38.4=py39h06a4308_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h5eee18b_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
